refactor: move the pipeline and worker into dial9-core
- pipeline contract (SegmentProcessor/SegmentData/ProcessError) -> core::pipeline, public
- worker engine -> core::worker behind a `pipeline` feature; default bus build pulls no runtime
- GzipCompressor/WriteBackProcessor -> core::worker::processors
- S3 upload -> core::worker::s3 behind `pipeline-s3`; telemetry's `worker-s3` forwards to it
