+++ dial9-core/Cargo.toml
@@ -16,20 +16,63 @@ crossbeam-queue = "0.3"
 dial9-trace-format = { workspace = true }
 libc = { workspace = true }
 tokio = { version = "1", features = ["sync"], default-features = false }
+tokio-util = { version = "0.7", optional = true }
+futures-util = { version = "0.3", default-features = false, features = [
+    "alloc",
+    "std",
+], optional = true }
+flate2 = { version = "1", optional = true }
 metrique = "0.1.23"
 metrique-timesource = "0.1"
 shuttle = { version = "0.9.1", optional = true }
 tracing = "0.1.44"
+aws-sdk-s3-transfer-manager = { version = "0.1.3", optional = true, default-features = false }
+aws-sdk-s3 = { version = "1", default-features = false, optional = true, features = [
+    "behavior-version-latest",
+    "sigv4a",
+    "rt-tokio",
+    "default-https-client",
+] }
+aws-config = { version = "1", optional = true, default-features = false }
+time = { version = "0.3", features = ["formatting", "macros"], optional = true }
+hostname = { version = "0.4", optional = true }
 
 [dev-dependencies]
 assert2 = { workspace = true }
+async-trait = "0.1.89"
 metrique-timesource = { version = "0.1", features = [
     "custom-timesource",
     "tokio",
 ] }
+metrique-writer = { version = "0.1", features = ["test-util"] }
 proptest = "1"
 tempfile = "3"
 tokio = { version = "1.52.0", features = ["rt", "macros", "time", "test-util"] }
+s3s = "0.13.0"
+s3s-fs = "0.13.0"
+s3s-aws = "0.13.0"
 
 [features]
 _shuttle = ["dep:shuttle"]
+## Exposes recording internals (`Batch`, `CentralCollector`, `encode_single`,
+## `drain_to_collector`) as `pub` for sibling-crate tests.
+test-util = []
+## The segment-processing pipeline runner: a background worker that drives a
+## `SegmentProcessor` chain over sealed segments. Pulls a tokio runtime.
+pipeline = [
+    "tokio/rt",
+    "tokio/time",
+    "tokio/macros",
+    "dep:tokio-util",
+    "dep:futures-util",
+    "dep:flate2",
+]
+## Built-in S3 upload stage. Implies the pipeline runner, pulls the aws SDK.
+pipeline-s3 = [
+    "pipeline",
+    "dep:aws-sdk-s3-transfer-manager",
+    "dep:aws-sdk-s3",
+    "dep:aws-config",
+    "dep:time",
+    "dep:hostname",
+]
