+++ dial9-core/src/worker/instance_metadata.rs
