+++ dial9-tokio-telemetry/src/lib.rs
@@ -12,7 +12,6 @@
 pub mod background_task;
 #[cfg(feature = "memory-profiling")]
 pub mod memory_profiling;
-pub(crate) mod metrics;
 pub(crate) use dial9_core::{primitives, rate_limit, sampling};
 #[cfg(feature = "taskdump")]
 pub(crate) mod task_dumped;
